feat: add validation errors to compatibility test in types package
