+++ lib/types/package.json
@@ -24,6 +24,7 @@
   },
   "dependencies": {
     "@dotcom-tool-kit/error": "^2.0.0",
+    "@dotcom-tool-kit/logger": "^2.0.0",
     "lodash.isplainobject": "^4.0.6",
     "lodash.mapvalues": "^4.6.0",
     "semver": "^7.3.7"
