+++ lib/types/src/index.ts
@@ -1,8 +1,9 @@
-import type { Logger } from 'winston'
-import { Schema, SchemaOutput } from './schema'
+import { styles as s } from '@dotcom-tool-kit/logger'
 import fs from 'fs'
 import path from 'path'
 import semver from 'semver'
+import type { Logger } from 'winston'
+import { Schema, SchemaOutput } from './schema'
 
 const packageJsonPath = path.resolve(__dirname, '../package.json')
 const packageJson = JSON.parse(fs.readFileSync(packageJsonPath, 'utf8'))
@@ -18,6 +19,67 @@ const typeSymbol = Symbol.for('@dotcom-tool-kit/types')
 const baseSymbol = Symbol.for('@dotcom-tool-kit/types/base')
 const taskSymbol = Symbol.for('@dotcom-tool-kit/types/task')
 const hookSymbol = Symbol.for('@dotcom-tool-kit/types/hook')
+
+export interface Invalid {
+  valid: false
+  reasons: string[]
+}
+export interface Valid<T> {
+  valid: true
+  value: T
+}
+export type Validated<T> = Invalid | Valid<T>
+
+export function mapValidated<T, U>(validated: Validated<T>, f: (val: T) => U): Validated<U> {
+  if (validated.valid) {
+    return { valid: true, value: f(validated.value) }
+  } else {
+    return validated
+  }
+}
+
+export function mapValidationError<T>(
+  validated: Validated<T>,
+  f: (reasons: string[]) => string[]
+): Validated<T> {
+  if (validated.valid) {
+    return validated
+  } else {
+    return { valid: false, reasons: f(validated.reasons) }
+  }
+}
+
+export function joinValidated<T, U>(first: Validated<T>, second: Validated<U>): Validated<[T, U]> {
+  if (first.valid) {
+    if (second.valid) {
+      return { valid: true, value: [first.value, second.value] }
+    } else {
+      return second
+    }
+  } else {
+    if (second.valid) {
+      return first
+    } else {
+      return { valid: false, reasons: [...first.reasons, ...second.reasons] }
+    }
+  }
+}
+
+export function reduceValidated<T>(validated: Validated<T>[]): Validated<T[]> {
+  let sequenced: Validated<T[]> = { valid: true, value: [] }
+  for (const val of validated) {
+    if (sequenced.valid) {
+      if (val.valid) {
+        sequenced.value.push(val.value)
+      } else {
+        sequenced = { valid: false, reasons: val.reasons }
+      }
+    } else if (!val.valid) {
+      sequenced.reasons.push(...val.reasons)
+    }
+  }
+  return sequenced
+}
 
 abstract class Base {
   static version = version
@@ -36,9 +98,18 @@ abstract class Base {
     return objectToCheck[typeSymbol] === this[typeSymbol]
   }
 
-  static isCompatible<T extends Base>(objectToCheck: unknown): objectToCheck is T {
+  static isCompatible<T extends Base>(objectToCheck: unknown): Validated<T> {
     if (!this.is(objectToCheck)) {
-      return false
+      return {
+        valid: false,
+        reasons: [
+          `${s.plugin(
+            '@dotcom-tool-kit/types'
+          )} type symbol is missing, make sure that this object derives from the ${s.code(
+            'Task'
+          )} or ${s.code('Hook')} class defined by the plugin`
+        ]
+      }
     }
 
     // objectToCheck from a plugin is compatible with this CLI if its version
@@ -49,7 +120,18 @@ abstract class Base {
     // this lets e.g. a CLI that includes types@2.2.0 load any plugin
     // that depends on any higher minor version of types.
     const range = `^${this.version}`
-    return semver.satisfies(objectToCheck.version, range)
+    if (semver.satisfies(objectToCheck.version, range)) {
+      return { valid: true, value: objectToCheck as T }
+    } else {
+      return {
+        valid: false,
+        reasons: [
+          `object is from an outdated version of ${s.plugin(
+            '@dotcom-tool-kit/types'
+          )}, make sure you're using at least version ${s.heading(this.version)} of the plugin`
+        ]
+      }
+    }
   }
 }
